Adds test for then second parameter
To meet the standard way of using promises in js the
readme specified that the then function would also
have a second parameter which would be treated the same
as if it was called with the error method. This was represented
in the code but not in the tests, the new test passes without
needing a change in the code.